+++ src/components/Button/index.jsx
@@ -3,8 +3,8 @@ const Button = ({ name, type, span }) => {
   return (
     <div
       className={`${styles.containerButtom} ${
-        type === "survey" ? styles["buttonSurvey"] : ""
-      } ${type === "submit" ? styles["buttonSubmit"] : ""}`}
+        type === "survey" ? styles.buttonSurvey : ""
+      } ${type === "submit" ? styles.buttonSubmit : ""}`}
     >
       <button type="button">
         <span>{span}</span>
