+++ src/components/ButtonsQuestions/ButtonQuestions.jsx
@@ -1,17 +1,7 @@
 import styles from "./style.module.css";
 
 const ContainerRadio = (props) => {
-  {
-    /*containerRadio
-  buttonRadio*/
-  }
   return (
-    // <div className={styles["containerRadio"]}>
-    //   <div className={styles["buttonRadio"]}>
-    //     <input type="radio"></input>
-    //   </div>
-    //   <label>{props.option}</label>
-    // </div>
     <div className={styles["containerRadio"]}>
       <label className={styles["buttonRadio"]}>
         <input type="radio" />
