--- src/components/SelectOptions/container.module.css
@@ -1,24 +1,0 @@
-.ContainerSelectOptions {
-  margin-bottom: 50px;
-}
-@media (min-width: 768px) {
-  .ContainerSelectOptions {
-    margin-bottom: 15px;
-    display: flex;
-  }
-  .ContainerSelectOptions label {
-    line-height: 60px;
-    margin-right: 10px;
-  }
-}
-
-@media (min-width: 992px) {
-  .ContainerSelectOptions {
-    margin: 0;
-    height: 30px;
-    margin-bottom: 15px;
-  }
-  .ContainerSelectOptions label {
-    line-height: 30px;
-  }
-}
