+++ src/components/SelectOptions/index.jsx
@@ -1,5 +1,4 @@
 import styles from "./style.module.css";
-import styles2 from "./container.module.css";
 import React, { useState, useEffect } from "react";
 import DropDownsOptions from "../DropDownsOptions";
 
@@ -11,23 +10,23 @@ const SelectOptions = (props) => {
   }, [props.placeHolder]);
   return (
     <>
-      <div className={`${styles2["ContainerSelectOptions"]}`}>
+      <div className={`${styles.containerSelectOptions}`}>
         <label>{props.title}</label>
         <div
           onClick={() => {
             active != "active" ? setactive("active") : setactive("");
           }}
-          className={`${styles["ContainerDropDown"]} ${
+          className={`${styles.containerDropDown} ${
             active != "" ? styles[active] : ""
           }`}
         >
           <input
             type="text"
-            className={styles["textBox"]}
+            className={styles.textBox}
             placeholder={placeHolder}
             readOnly
           />
-          <ul className={styles["option"]}>
+          <ul className={styles.option}>
             {props.options.map((item, key) => (
               <DropDownsOptions
                 key={key}
