+++ src/components/SelectOptions/style.module.css
@@ -1,15 +1,15 @@
-.ContainerSelectOptions {
+.containerSelectOptions {
   margin-bottom: 50px;
 }
 
-.ContainerDropDown {
+.containerDropDown {
   position: relative;
   width: 300px;
   height: 50px;
   border-bottom: 1px solid rgba(224, 90, 160, 1);
 }
 
-.ContainerDropDown::before {
+.containerDropDown::before {
   content: url(../../assets/images/Vector.svg);
   position: absolute;
   right: 20px;
@@ -19,12 +19,12 @@
   height: 8px;
   transition: 0.5s;
 }
-.ContainerDropDown.active::before {
+.containerDropDown.active::before {
   top: 25px;
   transform: rotate(-180deg);
 }
 
-.ContainerDropDown input {
+.containerDropDown input {
   position: absolute;
   top: 0;
   left: 0;
@@ -38,54 +38,54 @@
   padding: 12px 20px;
 }
 
-.ContainerDropDown .option {
+.containerDropDown .option {
   position: relative;
   top: 51px;
   background-color: #fff;
   overflow: hidden;
   display: none;
 }
-.ContainerDropDown.active .option {
+.containerDropDown.active .option {
   display: block;
 }
-.ContainerDropDown ul {
+.containerDropDown ul {
   list-style: none;
   padding-left: 0;
 }
 
-.ContainerDropDown li {
+.containerDropDown li {
   padding: 10px;
   border: 1px solid rgba(226, 228, 229, 1);
   cursor: pointer;
 }
 
-.ContainerDropDown li:hover {
+.containerDropDown li:hover {
   background-color: rgb(205, 200, 200);
 }
 
-.ContainerDropDown {
+.containerDropDown {
   font-size: 14px;
   font-weight: 400;
 }
 
 @media (min-width: 768px) {
-  .ContainerDropDown input {
+  .containerDropDown input {
     padding: 0px 12px;
     padding-top: 20px;
   }
 }
 @media (min-width: 992px) {
-  .ContainerDropDown {
+  .containerDropDown {
     height: 30px;
   }
-  .ContainerDropDown input {
+  .containerDropDown input {
     padding-top: 10px;
   }
-  .ContainerDropDown .option {
+  .containerDropDown .option {
     top: 31px;
     z-index: 1000000;
   }
-  .ContainerDropDown::before {
+  .containerDropDown::before {
     content: url(../../assets/images/Vector.svg);
     position: absolute;
     right: 20px;
@@ -94,5 +94,27 @@
     width: 3px;
     height: 3px;
     transition: 0.5s;
+  }
+}
+
+@media (min-width: 768px) {
+  .containerSelectOptions {
+    margin-bottom: 15px;
+    display: flex;
+  }
+  .containerSelectOptions label {
+    line-height: 60px;
+    margin-right: 10px;
+  }
+}
+
+@media (min-width: 992px) {
+  .containerSelectOptions {
+    margin: 0;
+    height: 30px;
+    margin-bottom: 15px;
+  }
+  .containerSelectOptions label {
+    line-height: 30px;
   }
 }
