+++ src/components/UnderlineQuestions/index.jsx
@@ -4,8 +4,8 @@ const UnderlineQuestions = (props) => {
   return (
     <div className={styles.containerQuestions}>
       <label>{props.question}</label>
-      {props.text === "años" ? (
-        <div className={styles["containerInput"]}>
+      {props.text ? (
+        <div className={styles.containerInput}>
           <input
             className={styles.inputAnswers}
             placeholder={[props.placeHolder != "" ? props.placeHolder : ""]}
